requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0

pypdf>=3.6.0
python-docx>=0.8.11

faiss-cpu>=1.7.4
numpy>=1.23.0

tqdm>=4.65.0

beautifulsoup4>=4.12.0
requests>=2.31.0

ddgs>=1.0.6

PyPDF2>=3.0.1

# If your code uses this (it appears imported in some versions)
lxml>=4.9.3
